Refs #5367. Fix Windows warnings.

+++ src/c/profile/transport/ip/tcp/tcp_transport_windows.c
@@ -3,6 +3,8 @@
 
 #include <uxr/client/util/time.h>
 
+#include <ws2tcpip.h>
+
 bool uxr_init_tcp_platform(
         struct uxrTCPPlatform* platform,
         uxrIpProtocol ip_protocol,
@@ -50,7 +52,7 @@ bool uxr_init_tcp_platform(
         {
             for (ptr = result; ptr != NULL; ptr = ptr->ai_next)
             {
-                if (0 == connect(platform->poll_fd.fd, ptr->ai_addr, ptr->ai_addrlen))
+                if (0 == connect(platform->poll_fd.fd, ptr->ai_addr, (int)ptr->ai_addrlen))
                 {
                     platform->poll_fd.events = POLLIN;
                     rv = true;
